Rename configuration files from tsx to ts
