+++ src/domain/registration/state/RegistrationSelectors.ts
